+++ CHANGELOG.md
@@ -95,6 +95,7 @@
 - Fixed instantly fleeing from Pokémon set to be unfleeable.
 - Fixed Pumpkaboo line forms not working. (Currently sizes aren't visual but check base stats to see which size you have.)
 - Fixed a bug that caused already interpreted messages for moves to be mistaken as uninterpreted.
+- Fixed a bug that would freeze a battle when a Pokémon gets trapped due to an ability.
 
 ### Developer
 - Add events that are fired just before and after a Pokémon is released (ReleasePokemonEvent.Pre and .Post)